requirements.txt
flask
ujson
pycrypto
jsonschema
configparser
pymysql
